+++ backend/src/main/java/com/handong/rebon/category/CategoryRepository.java
@@ -1,0 +1,6 @@
+package com.handong.rebon.category;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+}
