--- backend/src/main/java/com/handong/rebon/menu/domain/Menu.java
@@ -1,40 +1,0 @@
-package com.handong.rebon.menu.domain;
-
-import javax.persistence.*;
-
-import com.handong.rebon.shop.domain.Shop;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@Entity
-public class Menu {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-    private int price;
-
-    @ManyToOne
-    private MenuGroup menuGroup;
-
-    @ManyToOne
-    private Shop shop;
-
-    public Menu(String name, int price) {
-        this.name = name;
-        this.price = price;
-    }
-
-    public void belongTo(MenuGroup menuGroup) {
-        this.menuGroup = menuGroup;
-    }
-
-    public void belongShop(Shop shop) {
-        this.shop = shop;
-    }
-}
