--- backend/src/main/java/com/handong/rebon/menu/domain/MenuGroup.java
@@ -1,42 +1,0 @@
-package com.handong.rebon.menu.domain;
-
-import java.util.ArrayList;
-import java.util.List;
-import javax.persistence.*;
-
-import com.handong.rebon.shop.domain.Shop;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class MenuGroup {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Shop shop;
-
-    @OneToMany(mappedBy = "menuGroup", cascade = CascadeType.ALL)
-    private List<Menu> menus = new ArrayList<>();
-
-    public MenuGroup(String name) {
-        this.name = name;
-    }
-
-    public void addMenu(Menu menu) {
-        this.menus.add(menu);
-        menu.belongTo(this);
-    }
-
-    public void belongTo(Shop shop) {
-        this.shop = shop;
-    }
-}
