--- backend/src/main/java/com/handong/rebon/menu/domain/repository/MenuGroupRepository.java
@@ -1,8 +1,0 @@
-package com.handong.rebon.menu.domain.repository;
-
-import com.handong.rebon.menu.domain.MenuGroup;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
-}
