--- backend/src/main/java/com/handong/rebon/menu/presentation/dto/request/MenuRequest.java
@@ -1,14 +1,0 @@
-package com.handong.rebon.menu.presentation.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class MenuRequest {
-    private String name;
-    private int price;
-
-    public MenuRequest(String name, int price) {
-        this.name = name;
-        this.price = price;
-    }
-}
