+++ backend/src/main/java/com/handong/rebon/shop/application/adapter/LodgingServiceAdapter.java
@@ -1,8 +1,8 @@
 package com.handong.rebon.shop.application.adapter;
 
+import com.handong.rebon.category.Category;
 import com.handong.rebon.shop.application.dto.ShopRequestDto;
 import com.handong.rebon.shop.domain.Shop;
-import com.handong.rebon.shop.domain.category.Category;
 import com.handong.rebon.shop.domain.content.ShopContent;
 import com.handong.rebon.shop.domain.content.ShopImages;
 import com.handong.rebon.shop.domain.content.ShopScore;
