--- backend/src/main/java/com/handong/rebon/shop/domain/category/Categories.java
@@ -1,12 +1,0 @@
-package com.handong.rebon.shop.domain.category;
-
-import java.util.List;
-import javax.persistence.Embeddable;
-import javax.persistence.OneToMany;
-
-@Embeddable
-public class Categories {
-
-    @OneToMany(mappedBy = "parent")
-    private List<Category> categories;
-}
