--- backend/src/main/java/com/handong/rebon/shop/domain/category/CategoryRepository.java
@@ -1,6 +1,0 @@
-package com.handong.rebon.shop.domain.category;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
