+++ backend/src/main/java/com/handong/rebon/shop/domain/repository/MenuGroupRepository.java
@@ -1,0 +1,8 @@
+package com.handong.rebon.shop.domain.repository;
+
+import com.handong.rebon.shop.domain.menu.MenuGroup;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
+}
