--- backend/src/main/java/com/handong/rebon/shop/domain/tag/application/ShopTagService.java
@@ -1,7 +1,0 @@
-package com.handong.rebon.shop.domain.tag.application;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class ShopTagService {
-}
