--- backend/src/main/java/com/handong/rebon/shop/domain/tag/domain/repository/TagRepository.java
@@ -1,8 +1,0 @@
-package com.handong.rebon.shop.domain.tag.domain.repository;
-
-import com.handong.rebon.shop.domain.tag.domain.Tag;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TagRepository extends JpaRepository<Tag, Long> {
-}
