--- backend/src/main/java/com/handong/rebon/shop/domain/tag/presentation/AdminTagController.java
@@ -1,16 +1,0 @@
-package com.handong.rebon.shop.domain.tag.presentation;
-
-import com.handong.rebon.shop.domain.tag.application.TagService;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@RequestMapping("/admin/tags")
-@Controller
-public class AdminTagController {
-    private final TagService tagService;
-
-}
