+++ backend/src/main/java/com/handong/rebon/shop/domain/type/Lodging.java
@@ -3,8 +3,8 @@ package com.handong.rebon.shop.domain.type;
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
 
+import com.handong.rebon.category.Category;
 import com.handong.rebon.shop.domain.Shop;
-import com.handong.rebon.shop.domain.category.Category;
 import com.handong.rebon.shop.domain.content.ShopContent;
 import com.handong.rebon.shop.domain.content.ShopImages;
 import com.handong.rebon.shop.domain.content.ShopScore;
