+++ backend/src/main/java/com/handong/rebon/shop/presentation/dto/request/MenuRequest.java
@@ -1,0 +1,14 @@
+package com.handong.rebon.shop.presentation.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class MenuRequest {
+    private String name;
+    private int price;
+
+    public MenuRequest(String name, int price) {
+        this.name = name;
+        this.price = price;
+    }
+}
