+++ backend/src/main/java/com/handong/rebon/tag/domain/Tag.java
@@ -1,0 +1,34 @@
+package com.handong.rebon.tag.domain;
+
+import java.util.List;
+import javax.persistence.*;
+
+import com.handong.rebon.shop.domain.tag.ShopTag;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Tag {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    @OneToMany(mappedBy = "tag")
+    private List<ShopTag> shopTags;
+
+    public Tag(String name) {
+        this(null, name);
+    }
+
+    public Tag(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+}
