+++ backend/src/main/java/com/handong/rebon/tag/domain/repository/TagRepository.java
@@ -1,0 +1,8 @@
+package com.handong.rebon.tag.domain.repository;
+
+import com.handong.rebon.tag.domain.Tag;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TagRepository extends JpaRepository<Tag, Long> {
+}
