--- backend/src/test/java/com/handong/rebon/shop/domain/tag/domain/repository/TagRepositoryTest.java
@@ -1,5 +1,0 @@
-package com.handong.rebon.shop.domain.tag.domain.repository;
-
-class TagRepositoryTest {
-
-}
