+++ backend/src/test/java/com/handong/rebon/unit/shop/application/adapter/LodgingAdapterTest.java
@@ -1,9 +1,9 @@
 package com.handong.rebon.unit.shop.application.adapter;
 
+import com.handong.rebon.category.Category;
 import com.handong.rebon.shop.application.adapter.LodgingServiceAdapter;
 import com.handong.rebon.shop.application.dto.ShopRequestDto;
 import com.handong.rebon.shop.domain.Shop;
-import com.handong.rebon.shop.domain.category.Category;
 import com.handong.rebon.shop.domain.content.ShopImages;
 import com.handong.rebon.shop.domain.type.Lodging;
