+++ backend/src/test/java/com/handong/rebon/unit/shop/domain/ShopTest.java
@@ -4,14 +4,14 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
+import com.handong.rebon.category.Category;
 import com.handong.rebon.exception.shop.ShopTagNumberException;
-import com.handong.rebon.shop.domain.ShopCategory;
-import com.handong.rebon.shop.domain.category.Category;
+import com.handong.rebon.shop.domain.category.ShopCategory;
 import com.handong.rebon.shop.domain.content.ShopContent;
 import com.handong.rebon.shop.domain.content.ShopImages;
-import com.handong.rebon.shop.domain.tag.domain.ShopTag;
-import com.handong.rebon.shop.domain.tag.domain.Tag;
+import com.handong.rebon.shop.domain.tag.ShopTag;
 import com.handong.rebon.shop.domain.type.Restaurant;
+import com.handong.rebon.tag.domain.Tag;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
